Fix types, includes and compiler warnings
Clean up headers and types to silence warnings and improve correctness across the project:

- Reordered and normalized include directives in batventi.c and headers.
- Added pragma warning disables in btvenlib.h for WIP-unreferenced symbols.
- Replaced many int loop counters and length/count parameters with size_t where appropriate.
- Adjusted function signatures to use size_t (specifyParameter_multiple, loop indexes, etc.).
- Changed color-related types to UINT8 and updated related functions (findClosestConsoleColor, grayLevel) and structs.
- Fixed MultiByteToWideChar/MultiByteToWideChar size handling and cast to proper types.
- Initialized several variables and combined/normalized initializations to avoid uninitialized-use warnings.
- Minor API/behavior tweaks: changed WORD attr composition, switched WideCharToMultiByte return handling to size_t, small formatting/comment fixes.

These changes are intended to remove compiler warnings, use appropriate integer types for sizes/indices, and make the codebase more robust for 64-bit builds.

+++ batventi/batventi.c
@@ -1,9 +1,9 @@
 ﻿#define _CRT_SECURE_NO_WARNINGS
-#include <stdio.h>
-#include <Windows.h>
-#include <conio.h>
-#include <string.h>
 #include "resource.h"
+#include <conio.h>
+#include <stdio.h>
+#include <string.h>
+#include <Windows.h>
 
 #include "btvenlib.h"
 // btvenlib.h 这一行应该放到紧挨在标准库文件后面的位置，位于绝大多数语句之前
@@ -207,7 +207,7 @@ int main(int argc, char **argv)
 
 
 int handleargv1(const char funcName[]) {
-	for (int i = 0; commands[i].name != NULL; i++) {
+	for (size_t i = 0; commands[i].name != NULL; i++) {
 		if (_stricmp(funcName, commands[i].name) == 0) {
 			return commands[i].id;
 		}
